Added frontned login, home pages

+++ backend/utils/services.py
@@ -7,12 +7,16 @@ def reset_temp_folder():
     """
     removed_temp_files = []
 
-    for file in ('temp_path/external', 'temp_path/internal'):
+    for file in ('temp_files/external', 'temp_path/internal'):
         os.makedirs(file, exist_ok=True)
         for file_name in os.listdir(file):
             file_path = os.path.join(file, file_name)
-            removed_temp_files.append(str(file_path))
-            if os.path.isfile(file_path):
-                os.unlink(file_path)
+
+            if file_name.endswith('.gitkeep'):
+                continue
+            else:
+                removed_temp_files.append(str(file_path))
+                if os.path.isfile(file_path):
+                    os.unlink(file_path)
 
     return removed_temp_files
